// src/Models/settings.ts
export interface ISettings {
    id: string;
    shouldStop: number;
    pumpOnTime: number;
    pumpOffTime: number;
    pumpShouldStop: number;
    lightOnTime: number;
    lightOffTime: number;
    lightShouldStop: number;
    exhaustOnTime: number;
    exhaustOffTime: number;
    exhaustShouldStop: number;
    fanOnTime: number;
    fanOffTime: number;
    fanShouldStop: number;
}